pyproject.toml
[project]
name = "jobscraper"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "selenium (>=4.34.2,<5.0.0)",
    "logger (>=1.4,<2.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "jobscraper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
